logging states and some functionality fixes

+++ internal/domain/entity/ball.go
@@ -15,9 +15,9 @@ func (b *Ball) Fly(rightHand chan<- Ball) {
 	ticker := time.NewTicker(time.Second)
 	defer ticker.Stop()
 	for range ticker.C {
-		elapsed := int(time.Since(start).Seconds())
+		elapsed := int64(time.Since(start).Seconds())
 		fmt.Printf("Ball #%d: %d/%d seconds\n", b.ID, elapsed, b.Duration)
-		if elapsed >= int(b.Duration) {
+		if elapsed >= b.Duration {
 			rightHand <- *b
 			return
 		}
